% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccAFv2.R
\name{PredictCellCycle}
\alias{PredictCellCycle}
\title{Predict Cell Cycle}
\usage{
PredictCellCycle(seurat1, cutoff = 0.5)
}
\arguments{
\item{Seurat}{object that should have ccAFv2 cell cycle states predicted.}
}
\value{
Seurat object with ccAFv2 calls and probabilities for each cell cycle state.
}
\description{
This function predicts the cell cycle state for each cell in the object
using the ccAFv2 cell cycle classifier. \link{Which cell cycle states/phases}
\link{How to interpret data}
}
